backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
passlib==1.7.4
bcrypt==4.0.1
pyjwt
requests
cryptography
redis
